+++ src/main/java/lotto/study/Sum.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.study;
 
 import java.util.Arrays;
 import java.util.Objects;
