+++ src/test/java/lotto/LottoPurchaseTest.java
@@ -1,4 +1,4 @@
-package lotto.auto;
+package lotto;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
