+++ src/test/java/study/SumTest.java
@@ -1,5 +1,7 @@
-package lotto;
+package study;
 
+import lotto.study.StringParser;
+import lotto.study.Sum;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
